Rewrote most of the arm code
Added arm modes to simplify arm and grabber position. Changed the
buttons to match the reference sheet. To shoot, just press a shoot
button and it will go down, then shoot. Arm goes to hold by default.
Resets arm position in disabled so that if we disable with the arm down,
it will not go back down when we re-enable. Added in multiple shoot
modes for multiple distances but they have not been integrated with
their own values yet.